+++ cpp/mesh_gen.h
@@ -6,12 +6,12 @@ namespace tbem {
 
 template <size_t dim> struct Mesh;
 
-Mesh<3> sphere_mesh(const Vec3<double>& center, double radius, int refinements);
+Mesh<3> sphere_mesh(const Vec3<double>& center, double radius, size_t refinements);
 Mesh<3> rect_mesh(const Vec3<double>& lower_left, const Vec3<double>& upper_left,
     const Vec3<double>& upper_right, const Vec3<double>& lower_right);
 
 Mesh<2> line_mesh(const Vec2<double>& a, const Vec2<double>& b);
-Mesh<2> circle_mesh(const Vec2<double>& center, double r, int refinements);
+Mesh<2> circle_mesh(const Vec2<double>& center, double r, size_t refinements);
 
 } //END NAMESPACE tbem
